create free

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: rgatnaou <rgatnaou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/09 16:36:14 by rgatnaou          #+#    #+#             */
-/*   Updated: 2026/01/09 16:43:29 by rgatnaou         ###   ########.fr       */
+/*   Updated: 2026/01/10 19:44:46 by rgatnaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 #include <stddef.h>
 
-int		ft_strlen(const char *s);
+size_t	ft_strlen(const char *s);
 void	*ft_memset(void *b, int c, size_t len);
